% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_brand_caption.R
\name{create_brand_caption}
\alias{create_brand_caption}
\title{Title create brand caption}
\usage{
create_brand_caption(
  brand_icon = "f09b",
  caption_text,
  caption_text_font = "sans"
)
}
\arguments{
\item{brand_icon}{The Fonts Awesome Brand icon code to use. Default is "f09b" (GitHub Icon).}

\item{caption_text}{The text to use in the caption.}

\item{caption_text_font}{The font to use for the caption text. Default is "sans".}
}
\value{
A ggplot2::labs caption with the specified brand icon and caption text.
}
\description{
This function allows you to create a brand caption for ggplot2 plots, using Font Awesome 6 Brands Regular.
}
\examples{
library(ggplot2)
library(showtext)
library(ggtext)
library(sysfonts)
ggplot() +
create_brand_caption(caption_text = "ggcolrr") +
theme(plot.caption = ggtext::element_textbox_simple(margin = margin(3,0,0,0,"mm"), size = 15))
}
