% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_light.R
\name{main_light}
\alias{main_light}
\title{A light theme for ggplot2}
\usage{
main_light(..., base_size = 12)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::theme}.}

\item{base_size}{Numeric. The base size for text elements. Default is 12.}
}
\value{
A \code{ggplot2::theme} object with customized text sizes and colors.
}
\description{
This function creates a ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  main_light(base_size = 14)
}
