% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_theme.R
\name{scale_color_theme}
\alias{scale_color_theme}
\title{Title theme scale color}
\usage{
scale_color_theme(
  theme = "nord_aurora",
  theme_key = list(nord_polar_night = c("#2E3440", "#3B4252", "#434C5E", "#4C566A"),
    nord_snow_storm = c("#D8DEE9", "#E5E9F0", "#ECEFF4"), nord_frost = c("#8FBCBB",
    "#88C0D0", "#81A1C1", "#5E81AC"), nord_aurora = c("#BF616A", "#D08770", "#EBCB8B",
    "#A3BE8C", "#B48EAD"))
)
}
\arguments{
\item{theme}{Character. The theme to use. Default is "nord_aurora".}

\item{theme_key}{List. A list of color values for the theme. Default is a list of Nord colors.}
}
\value{
'scale_color_manual' with the specified theme.
}
\description{
This function allows you to set manual color scales for ggplot2 plots.
}
\examples{
library(ggplot2)
mtcars$cyl <- as.factor(mtcars$cyl)
ggplot(data = mtcars, aes(x = hp, y = mpg, color = cyl, group = cyl)) +
 geom_line() +
 scale_color_theme(theme = "nord_aurora")
}
