% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nord.R
\name{theme_nord}
\alias{theme_nord}
\title{A Nord theme for ggplot2}
\usage{
theme_nord(..., light = TRUE, base_size = 12, font = "Open Sans")
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::theme}.}

\item{base_size}{Numeric. The base size for text elements. Default is 12.}
}
\value{
A \code{ggplot2::theme} object with customized text sizes and colors.
}
\description{
This function creates a ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_nord(light = TRUE, base_size = 14, font = "Open Sans")
}
